# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MusicGetter Extension API",
    "version": "1.0.0",
    "description": "Rendered metadata only. No Spotify/Yandex/VK credentials, cookies, source auth headers, HTML or extra JSON properties. Bearer means MusicGetter token only. JSON property names are case-sensitive; duplicates rejected. All responses no-store. CORS allows only exact configured chrome-extension/moz-extension origins; no Origin is permitted for extension service workers/native tools with Bearer. No cookie authentication. Limits are per-process fixed-minute windows: IP 120, owner 60, pair claim IP 5; configurable via environment. Retry-After on 429. Reverse proxy headers are not trusted."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development only; production must use HTTPS."
    }
  ],
  "security": [
    {
      "extensionBearer": []
    }
  ],
  "paths": {
    "/v1/pair/claim": {
      "post": {
        "operationId": "claimPairing",
        "description": "Consume a bot-issued code and issue a 30-day MusicGetter session. Token is returned once. Lost response requires a new /connect; no replay can retrieve the token. No Authorization header on this endpoint.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          }
        },
        "x-max-body-bytes": 1024,
        "security": []
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "description": "Current user and authenticated extension session. No request body.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          }
        }
      }
    },
    "/v1/destinations": {
      "get": {
        "operationId": "listDestinations",
        "description": "Owner-scoped saved destination collections; no calls to a music bot. Empty until a destination is provisioned. UUID keyset; next_cursor may lead to an empty final page.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinations"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    },
    "/v1/imports": {
      "post": {
        "operationId": "createImport",
        "description": "Create collecting import and find/create source profile/collection for current user. client_request_id is unique per user; canonical typed payload is immutable across retries. Destination must belong to current user.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImport"
              },
              "example": {
                "client_request_id": "create-unique-client-key",
                "source": {
                  "service": "spotify",
                  "profile_key": "personal",
                  "collection_key": "favorites",
                  "kind": "favorites",
                  "title": "Liked songs"
                },
                "destination_collection_id": "00000000-0000-0000-0000-000000000001"
              }
            }
          }
        },
        "x-max-body-bytes": 16384,
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "If supplied, must equal client_request_id (create) or idempotency_key (tracks)."
          }
        ]
      }
    },
...
    "/v1/pairings/redeem": {
      "post": {
        "operationId": "legacyRedeem",
        "description": "Deprecated alias of /v1/pair/claim with previous response shape; shares rate limits and validation.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyClaimResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          }
        },
        "x-max-body-bytes": 1024,
        "security": [],
        "deprecated": true
      }
    },
    "/v1/extension/session": {
      "get": {
        "operationId": "legacySession",
        "description": "Deprecated alias: authenticated session metadata.",
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "description": "Safe correlation ID, also included in errors.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacySession"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    }
  },
  "components": {
    "securitySchemes": {
      "extensionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "mge_<256-bit base64url>",
        "description": "Own MusicGetter session token, never a streaming credential."
      }
    },
    "schemas": {
...
      "ClaimRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z2-7]{26}$",
            "description": "One-time 128-bit Base32 code from /connect. Case-insensitive, surrounding whitespace trimmed by server."
          }
        },
        "required": [
          "code"
        ]
      },
      "ClaimResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^mge_[A-Za-z0-9_-]{43}$"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "token",
          "token_type",
          "expires_at"
        ]
      },
      "Me": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          },
          "telegram_user_id": {
            "type": "integer",
            "minimum": 1
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          },
          "session_expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "telegram_user_id",
          "session_id",
          "session_expires_at"
        ]
      },
...
      "Destinations": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Destination"
            }
          },
          "next_cursor": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          }
        },
        "required": [
          "items"
        ]
      },
...
      "Created": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          },
          "replay": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "replay"
        ]
      },
...
      "LegacyClaimResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "session": {
            "$ref": "#/components/schemas/LegacySession"
          }
        },
        "required": [
          "token",
          "tokenType",
          "session"
        ]
      }
    },
    "responses": {
      "Error400": {
        "description": "Malformed/duplicate/unknown JSON fields, unsupported query or credentials.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing, invalid, expired or revoked MusicGetter credentials.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error403": {
        "description": "Origin or CORS preflight rejected.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error404": {
        "description": "Unknown resource or resource owned by another user.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error405": {
        "description": "Method not allowed.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error409": {
        "description": "Changed idempotency payload/key, import state or incomplete chunk range.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error413": {
        "description": "Body too large.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error415": {
        "description": "Use uncompressed UTF-8 application/json.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error422": {
        "description": "Invalid fields, counts or import quota exceeded.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error429": {
        "description": "Rate limit exceeded; honor Retry-After.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          },
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "Delay in seconds before retry."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error503": {
        "description": "Storage unavailable or timed out. Retry mutation using the same key/body.",
        "headers": {
          "X-Request-ID": {
            "description": "Safe correlation ID, also included in errors.",
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "no-store"
              ]
            }
          },
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "Delay in seconds before retry."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}